conf/distro: cml-debug-tiny: introduce feature 'tmedbg-extended'
Added "tmedbg-extended" to DISTRO_FEATURES which is used to
select extended packages for cml-initramfs. Currently this is gdb
and valgrind which are extremely large in binary size.

Layers for Embedded boards such as the rpi can than easily be
configured to remove this feature, to not exceed kernel+initrd
binary size for the corresponding bootloader.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>